+++ src/rolling/builder.rs
@@ -65,7 +65,7 @@ impl Builder {
     ///
     /// ```
     /// # fn docs() {
-    /// use tracing_appender::rolling::{Rotation, RollingFileAppender};
+    /// use tracing_appender_plus::rolling::{Rotation, RollingFileAppender};
     ///
     /// let appender = RollingFileAppender::builder()
     ///     .rotation(Rotation::HOURLY) // rotate log files once every hour
@@ -94,7 +94,7 @@ impl Builder {
     /// Setting a prefix:
     ///
     /// ```
-    /// use tracing_appender::rolling::RollingFileAppender;
+    /// use tracing_appender_plus::rolling::RollingFileAppender;
     ///
     /// # fn docs() {
     /// let appender = RollingFileAppender::builder()
@@ -109,7 +109,7 @@ impl Builder {
     /// No prefix:
     ///
     /// ```
-    /// use tracing_appender::rolling::RollingFileAppender;
+    /// use tracing_appender_plus::rolling::RollingFileAppender;
     ///
     /// # fn docs() {
     /// let appender = RollingFileAppender::builder()
@@ -146,7 +146,7 @@ impl Builder {
     /// Setting a suffix:
     ///
     /// ```
-    /// use tracing_appender::rolling::RollingFileAppender;
+    /// use tracing_appender_plus::rolling::RollingFileAppender;
     ///
     /// # fn docs() {
     /// let appender = RollingFileAppender::builder()
@@ -161,7 +161,7 @@ impl Builder {
     /// No suffix:
     ///
     /// ```
-    /// use tracing_appender::rolling::RollingFileAppender;
+    /// use tracing_appender_plus::rolling::RollingFileAppender;
     ///
     /// # fn docs() {
     /// let appender = RollingFileAppender::builder()
@@ -189,9 +189,13 @@ impl Builder {
 
     /// Keeps the last `n` log files on disk.
     ///
-    /// When a new log file is created, if there are `n` or more
-    /// existing log files in the directory, the oldest will be deleted.
-    /// If no value is supplied, the `RollingAppender` will not remove any files.
+    /// When constructing a `RollingFileAppender` or starting a new log file,
+    /// the appender will delete the oldest matching log files until at most `n`
+    /// files remain. The exact number of retained files can sometimes dip below
+    /// the maximum, so if you need to retain `m` log files, specify a max of
+    /// `m + 1`.
+    ///
+    /// If `0` is supplied, the `RollingAppender` will not remove any files.
     ///
     /// Files are considered candidates for deletion based on the following
     /// criteria:
@@ -214,7 +218,7 @@ impl Builder {
     /// # Examples
     ///
     /// ```
-    /// use tracing_appender::rolling::RollingFileAppender;
+    /// use tracing_appender_plus::rolling::RollingFileAppender;
     ///
     /// # fn docs() {
     /// let appender = RollingFileAppender::builder()
@@ -228,7 +232,8 @@ impl Builder {
     #[must_use]
     pub fn max_log_files(self, n: usize) -> Self {
         Self {
-            max_files: Some(n),
+            // Setting `n` to 0 will disable the max files (effectively make it infinite).
+            max_files: Some(n).filter(|&n| n > 0),
             ..self
         }
     }
@@ -242,7 +247,7 @@ impl Builder {
     /// # Examples
     ///
     /// ```
-    /// use tracing_appender::rolling::{Rotation, RollingFileAppender};
+    /// use tracing_appender_plus::rolling::{Rotation, RollingFileAppender};
     ///
     /// # fn docs() {
     /// let appender = RollingFileAppender::builder()
@@ -257,7 +262,7 @@ impl Builder {
     /// This is equivalent to
     /// ```
     /// # fn docs() {
-    /// let appender = tracing_appender::rolling::daily("myapp.log", "/var/log/myapp");
+    /// let appender = tracing_appender_plus::rolling::daily("myapp.log", "/var/log/myapp");
     /// # drop(appender);
     /// # }
     /// ```